stable but nonfunctional TOS componant
